requirements.txt
pandas~=1.5.0
matplotlib~=3.6.0
seaborn~=0.12.0
numpy~=1.23.3
statsmodels~=0.13.2
sentinelhub~=3.7.0
opencv-python~=4.6.0.66
Pillow~=9.2.0
rasterio~=1.3.4
beautifulsoup4~=4.11.1
pygam~=0.8.0